removed resources from ui
